+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.16.1
  * Add an analysis_options.yaml and fix or suppress all the complaints from it.
+ * Add unit tests using dart:ffi to compare formatting output to ICU.
+ * Bump SDK requirements up to 2.5.0 for dart:ffi availability.
 
 ## 0.16.0
  * Fix 'k' formatting (1 to 24 hours) which incorrectly showed 0 to 23.